requirements.txt
torch==2.0.0
onnxruntime-gpu==1.12.0
opencv-python==4.7.0.72
numpy==1.26.4
tqdm==4.66.5